Refactor requestHandler Function and add POST handler
