+++ deps.ts
@@ -1,2 +1,5 @@
 export { serve } from "https://deno.land/std@0.125.0/http/server.ts";
-export { readableStreamFromReader } from "https://deno.land/std@0.125.0/streams/mod.ts";
+export {
+	readableStreamFromReader,
+	writableStreamFromWriter,
+} from "https://deno.land/std@0.125.0/streams/mod.ts";
